🚚️ api: move UnknownCountryForStudentEnrolmentError to src/enrolment context
