+++ api/lib/domain/errors.js
@@ -754,15 +754,6 @@ class UserNotFoundError extends NotFoundError {
   }
 }
 
-class UnknownCountryForStudentEnrolmentError extends DomainError {
-  constructor(
-    { firstName, lastName },
-    message = `L'élève ${firstName} ${lastName} a été inscrit avec un code pays de naissance invalide. Veuillez corriger ses informations sur l'espace PixOrga de l'établissement ou contacter le support Pix`,
-  ) {
-    super(message);
-  }
-}
-
 class WrongDateFormatError extends DomainError {
   constructor(message = 'Format de date invalide.') {
     super(message);
@@ -1042,7 +1033,6 @@ export {
   TargetProfileInvalidError,
   TooManyRows,
   UnexpectedUserAccountError,
-  UnknownCountryForStudentEnrolmentError,
   UserAlreadyExistsWithAuthenticationMethodError,
   UserAlreadyLinkedToCandidateInSessionError,
   UserCouldNotBeReconciledError,
