+++ api/src/certification/enrolment/domain/errors.js
@@ -18,8 +18,18 @@ class CertificationCandidateNotFoundError extends DomainError {
   }
 }
 
+class UnknownCountryForStudentEnrolmentError extends DomainError {
+  constructor(
+    { firstName, lastName },
+    message = `L'élève ${firstName} ${lastName} a été inscrit avec un code pays de naissance invalide. Veuillez corriger ses informations sur l'espace PixOrga de l'établissement ou contacter le support Pix`,
+  ) {
+    super(message);
+  }
+}
+
 export {
   CertificationCandidateForbiddenDeletionError,
   CertificationCandidateNotFoundError,
   SessionStartedDeletionError,
+  UnknownCountryForStudentEnrolmentError,
 };
