+++ api/tests/integration/application/error-manager_test.js
@@ -666,18 +666,6 @@ describe('Integration | API | Controller Error', function () {
       expect(response.statusCode).to.equal(UNPROCESSABLE_ENTITY_ERROR);
     });
 
-    it('responds UnprocessableEntity when a UnknownCountryForStudentEnrolmentError occurs', async function () {
-      routeHandler.throws(
-        new DomainErrors.UnknownCountryForStudentEnrolmentError({ firstName: 'Paul', lastName: 'Preboist' }),
-      );
-      const response = await server.requestObject(request);
-
-      expect(response.statusCode).to.equal(UNPROCESSABLE_ENTITY_ERROR);
-      expect(responseDetail(response)).to.equal(
-        "L'élève Paul Preboist a été inscrit avec un code pays de naissance invalide. Veuillez corriger ses informations sur l'espace PixOrga de l'établissement ou contacter le support Pix",
-      );
-    });
-
     it('responds UnprocessableEntity when a AuthenticationMethodAlreadyExistsError occurs', async function () {
       routeHandler.throws(new DomainErrors.AuthenticationMethodAlreadyExistsError());
       const response = await server.requestObject(request);
